Extract a standardized node with simple refactoring
