+++ ui/src/components/branch-display/index.ts
@@ -1,0 +1,2 @@
+export type * from './types';
+export { BranchSvgCircle } from './BranchSvgCircle';
